requirements.txt
Flask==2.0.1
python-decouple==3.5
requests==2.26.0
gunicorn==20.1.0